Refactor: extract specs2 macro methods to scala-2
